A little bit optimizations and portable Makefile
This is final version, I think. The architecture of LuaJIT is adverse for e2k.  Random access to BC templates and all operands from memory slows down compiler.
To make optimal lua jit compiler for e2k, should be done template-chain architecture and most data in registers with JIT planning of operations and codegen. But within the framework of this project, it is not advisable